// Neptune.Web/src/app/shared/generated/model/person-detail-dto.ts
export class PersonDetailDto { 
    PersonID?: number;
    FirstName?: string | null;
    LastName?: string | null;
    Email?: string | null;
    Phone?: string | null;
    CreateDate?: string;
    UpdateDate?: string | null;
    LastActivityDate?: string | null;
    IsActive?: boolean;
    RoleID?: number;
    RoleName?: string | null;
    RoleDisplayName?: string | null;
    Organization?: OrganizationSimpleDto;
    IsOCTAGrantReviewer?: boolean;
    ReceiveSupportEmails?: boolean;
    ReceiveRSBRevisionRequestEmails?: boolean;
    AssignedStormwaterJurisdictions?: Array<StormwaterJurisdictionDisplayDto> | null;
    PrimaryContactOrganizations?: Array<OrganizationSimpleDto> | null;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}